fixed for crashesPerMillion miles -- was sometimes dividing by 0 and get NaN because of using longValue() instead of doubleValue

+++ dao/src/main/java/com/inthinc/pro/model/CrashSummary.java
@@ -77,7 +77,7 @@ public class CrashSummary extends BaseEntity {
 			crashesPerMillionMiles = 0.0;
 			return;
 		}
-		crashesPerMillionMiles = (new Double(crashesInTimePeriod)*1000000.0)/new Double(totalMiles.longValue());
+		crashesPerMillionMiles = (new Double(crashesInTimePeriod)*1000000.0)/new Double(totalMiles.doubleValue());
 	}
 
 	public Integer getCrashesInTimePeriod() {
